requirements.txt
streamlit
pandas
numpy
scipy
statsmodels
pingouin
plotly
openpyxl
python-docx
